requirements.txt
fastapi
uvicorn[standard]
rembg
pillow
opencv-python-headless
numpy